feat employees/credentials/positions

+++ NarzedzieHR/NarzedzieHR/Service/StanowiskoService.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
+using System.Windows.Forms;
 
 namespace NarzedzieHR.Service
 {
@@ -10,9 +11,9 @@ namespace NarzedzieHR.Service
     {
         private readonly string _connectionString = "data source=sql.bsite.net\\MSSQL2016;initial catalog=kapi1023_;user id=kapi1023_;password=Haslo123#$";
 
-        public DataTable GetAllStanowiska()
+        public DataSet GetAllStanowiska()
         {
-            DataTable dataTable = new DataTable();
+            DataSet dataSet = new DataSet();
 
             using (SqlConnection connection = new SqlConnection(_connectionString))
             {
@@ -22,7 +23,7 @@ namespace NarzedzieHR.Service
                 try
                 {
                     connection.Open();
-                    dataAdapter.Fill(dataTable);
+                    dataAdapter.Fill(dataSet);
                 }
                 catch (Exception ex)
                 {
@@ -30,7 +31,8 @@ namespace NarzedzieHR.Service
                 }
             }
 
-            return dataTable;
+        
+            return dataSet;
         }
 
         public bool AddStanowisko(StanowiskoModel stanowisko)
@@ -40,16 +42,25 @@ namespace NarzedzieHR.Service
                 try
                 {
                     connection.Open();
-                    SqlDataAdapter adapter = new SqlDataAdapter();
-                    SqlCommand command = new SqlCommand("INSERT INTO Stanowisko (Nazwa, Opis, DzialId, StawkaWynagrodzenia) VALUES (@Nazwa, @Opis, @DzialId, @StawkaWynagrodzenia)", connection);
+                    SqlCommand command = new SqlCommand("INSERT INTO Stanowisko (Nazwa, Opis, DzialId, StawkaWynagrodzenia) VALUES (@Nazwa, @Opis, @DzialId, @StawkaWynagrodzenia); SELECT SCOPE_IDENTITY();", connection);
                     command.Parameters.AddWithValue("@Nazwa", stanowisko.Nazwa);
                     command.Parameters.AddWithValue("@Opis", stanowisko.Opis);
                     command.Parameters.AddWithValue("@DzialId", stanowisko.DzialId);
                     command.Parameters.AddWithValue("@StawkaWynagrodzenia", stanowisko.StawkaWynagrodzenia);
-                    adapter.InsertCommand = command;
-                    int rowsAffected = adapter.InsertCommand.ExecuteNonQuery();
 
-                    return rowsAffected > 0;
+                    int insertedStanowiskoId = Convert.ToInt32(command.ExecuteScalar());
+                    if (stanowisko.Benefits != null)
+                    {
+                        foreach (var benefit in stanowisko.Benefits)
+                        {
+                            SqlCommand addBenefitCommand = new SqlCommand("INSERT INTO BenefitStanowisko (BenefitId, StanowiskoId) VALUES (@BenefitId, @StanowiskoId)", connection);
+                            addBenefitCommand.Parameters.AddWithValue("@BenefitId", benefit.Id);
+                            addBenefitCommand.Parameters.AddWithValue("@StanowiskoId", insertedStanowiskoId);
+                            addBenefitCommand.ExecuteNonQuery();
+                        }
+                    }
+
+                    return insertedStanowiskoId > 0;
                 }
                 catch (Exception ex)
                 {
@@ -66,25 +77,52 @@ namespace NarzedzieHR.Service
                 try
                 {
                     connection.Open();
-                    SqlDataAdapter adapter = new SqlDataAdapter();
-                    SqlCommand command = new SqlCommand("UPDATE Stanowisko SET Nazwa = @Nazwa, Opis = @Opis, DzialId = @DzialId, StawkaWynagrodzenia = @StawkaWynagrodzenia WHERE Id = @Id", connection);
-                    command.Parameters.AddWithValue("@Nazwa", stanowisko.Nazwa);
-                    command.Parameters.AddWithValue("@Opis", stanowisko.Opis);
-                    command.Parameters.AddWithValue("@DzialId", stanowisko.DzialId);
-                    command.Parameters.AddWithValue("@StawkaWynagrodzenia", stanowisko.StawkaWynagrodzenia);
-                    command.Parameters.AddWithValue("@Id", stanowisko.Id);
-                    adapter.UpdateCommand = command;
-                    int rowsAffected = adapter.UpdateCommand.ExecuteNonQuery();
 
-                    return rowsAffected > 0;
+                    SqlCommand updateCommand = new SqlCommand("UPDATE Stanowisko SET Nazwa = @Nazwa, Opis = @Opis, DzialId = @DzialId, StawkaWynagrodzenia = @StawkaWynagrodzenia WHERE Id = @Id", connection);
+                    updateCommand.Parameters.AddWithValue("@Nazwa", stanowisko.Nazwa);
+                    updateCommand.Parameters.AddWithValue("@Opis", stanowisko.Opis);
+                    updateCommand.Parameters.AddWithValue("@DzialId", stanowisko.DzialId);
+                    updateCommand.Parameters.AddWithValue("@StawkaWynagrodzenia", stanowisko.StawkaWynagrodzenia);
+                    updateCommand.Parameters.AddWithValue("@Id", stanowisko.Id);
+
+                    int rowsUpdated = updateCommand.ExecuteNonQuery();
+
+                    if (rowsUpdated > 0)
+                    {
+                        SqlCommand deleteBenefitsCommand = new SqlCommand("DELETE FROM BenefitStanowisko WHERE StanowiskoId = @StanowiskoId", connection);
+                        deleteBenefitsCommand.Parameters.AddWithValue("@StanowiskoId", stanowisko.Id);
+                        deleteBenefitsCommand.ExecuteNonQuery();
+
+                        if (stanowisko.Benefits != null && stanowisko.Benefits.Count > 0)
+                        {
+                            SqlCommand insertBenefitCommand = new SqlCommand("INSERT INTO BenefitStanowisko (BenefitId, StanowiskoId) VALUES (@BenefitId, @StanowiskoId)", connection);
+
+                            insertBenefitCommand.Parameters.Add("@BenefitId", SqlDbType.Int);
+                            insertBenefitCommand.Parameters.Add("@StanowiskoId", SqlDbType.Int);
+
+                            foreach (var benefit in stanowisko.Benefits)
+                            {
+                                insertBenefitCommand.Parameters["@BenefitId"].Value = benefit.Id;
+                                insertBenefitCommand.Parameters["@StanowiskoId"].Value = stanowisko.Id;
+                                insertBenefitCommand.ExecuteNonQuery();
+                            }
+                        }
+
+                        return true;
+                    }
+
+                        return false; 
+                    
                 }
                 catch (Exception ex)
                 {
                     Console.WriteLine(ex.Message);
-                    return false;
+                    return false; // Return false if an exception occurred
                 }
             }
         }
+
+
 
         public int GetDzialIdForStanowisko(int stanowiskoId)
         {
@@ -122,10 +160,27 @@ namespace NarzedzieHR.Service
                 try
                 {
                     connection.Open();
+
                     SqlDataAdapter adapter = new SqlDataAdapter();
+                    SqlCommand checkBenefitCommand = new SqlCommand("SELECT COUNT(*) FROM BenefitStanowisko WHERE StanowiskoId = @StanowiskoId", connection);
+                    checkBenefitCommand.Parameters.AddWithValue("@StanowiskoId", stanowiskoId);
+                    adapter.SelectCommand = checkBenefitCommand;
+
+                    DataSet dataSet = new DataSet();
+                    adapter.Fill(dataSet);
+
+                    int benefitCount = Convert.ToInt32(dataSet.Tables[0].Rows[0][0]);
+
+                    if (benefitCount > 0)
+                    {
+                        MessageBox.Show("Nie można usunąć stanowiska, które ma przypisane benefity.");
+                        return false; 
+                    }
+
                     SqlCommand deleteCommand = new SqlCommand("DELETE FROM Stanowisko WHERE Id = @Id", connection);
                     deleteCommand.Parameters.AddWithValue("@Id", stanowiskoId);
                     adapter.DeleteCommand = deleteCommand;
+
                     int rowsAffected = adapter.DeleteCommand.ExecuteNonQuery();
 
                     return rowsAffected > 0;
@@ -135,7 +190,36 @@ namespace NarzedzieHR.Service
                     Console.WriteLine(ex.Message);
                     return false;
                 }
+                finally
+                {
+                    connection.Close();
+                }
             }
+        }
+
+
+        public DataSet GetBenefitsForStanowisko(int stanowiskoId)
+        {
+            DataSet dataSet = new DataSet();
+
+            using (SqlConnection connection = new SqlConnection(_connectionString))
+            {
+                SqlDataAdapter dataAdapter = new SqlDataAdapter();
+                dataAdapter.SelectCommand = new SqlCommand("SELECT b.*, CASE WHEN bs.StanowiskoId IS NOT NULL THEN 1 ELSE 0 END AS czyWybrana FROM Benefit b LEFT JOIN BenefitStanowisko bs ON b.Id = bs.BenefitId AND bs.StanowiskoId = @StanowiskoId", connection);
+                dataAdapter.SelectCommand.Parameters.AddWithValue("@StanowiskoId", stanowiskoId);
+
+                try
+                {
+                    connection.Open();
+                    dataAdapter.Fill(dataSet);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
+            }
+
+            return dataSet;
         }
     }
 }
